+++ object-utils/documentation/routes/home-element/examples/deep-freeze/index.js
@@ -1,0 +1,3 @@
+export {default as jsCode} from '!raw-loader!./javascript';
+export {default as templateCode} from '!raw-loader!./template';
+export {default as cssCode} from '!raw-loader!./index.css';