+++ object-utils/documentation/routes/home-element/examples/deep-freeze/template.js
@@ -1,0 +1,5 @@
+(context, html) => html`
+<pre>${context.error}</pre>
+
+<pre>user.account.active = ${context.user.account.active}</pre>
+`;